# docs/swagger.yaml
definitions:
...
  client.ClientResponse:
    properties:
      allowed_scopes:
        items:
          type: string
        type: array
      client_id:
        type: string
      client_name:
        type: string
      client_secret:
        type: string
      client_type:
        type: string
      grant_types:
        items:
          type: string
        type: array
      redirect_uris:
        items:
          type: string
        type: array
    type: object
  client.CreateClientRequest:
    properties:
      allowed_scopes:
        items:
          type: string
        type: array
      client_id:
        description: optional; auto-generated if empty
        type: string
      client_name:
        type: string
      client_secret:
        description: optional; auto-generated if empty (confidential only)
        type: string
      client_type:
        type: string
      grant_types:
        items:
          type: string
        type: array
      redirect_uris:
        items:
          type: string
        type: array
    type: object
  client.UpdateClientRequest:
    properties:
      allowed_scopes:
        items:
          type: string
        type: array
      client_name:
        type: string
      redirect_uris:
        items:
          type: string
        type: array
    type: object
  httputil.Error:
    properties:
      code:
        type: string
      details:
        additionalProperties: {}
        type: object
      message:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /.well-known/openid-configuration:
    get:
      description: Returns the OpenID Connect discovery document (RFC 8414).
      parameters:
      - description: Tenant identifier
        in: header
        name: X-Tenant-ID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: OIDC discovery document
      tags:
      - oidc
...
  /device/authorize:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Initiate the OAuth 2.0 Device Authorization Grant (RFC 8628). Poll
        /token with grant_type=urn:ietf:params:oauth:grant-type:device_code.
      parameters:
      - description: Tenant identifier
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Client ID
        in: formData
        name: client_id
        required: true
        type: string
      - description: Requested scopes
        in: formData
        name: scope
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Error'
      summary: Device authorization
      tags:
      - oidc
...
  /tenants/{tid}/clients:
    get:
      parameters:
      - description: Tenant ID
        in: path
        name: tid
        required: true
        type: string
      - description: Pagination offset
        in: query
        name: offset
        type: integer
      - description: Page size (default 20)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/client.ClientResponse'
            type: array
      security:
      - AdminAuth: []
      summary: List clients
      tags:
      - clients
    post:
      consumes:
      - application/json
      parameters:
      - description: Tenant ID
        in: path
        name: tid
        required: true
        type: string
      - description: Client definition
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/client.CreateClientRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/client.ClientResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Error'
      security:
      - AdminAuth: []
      summary: Create client
      tags:
      - clients
  /tenants/{tid}/clients/{cid}:
    delete:
      parameters:
      - description: Tenant ID
        in: path
        name: tid
        required: true
        type: string
      - description: Client ID
        in: path
        name: cid
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Error'
      security:
      - AdminAuth: []
      summary: Delete client
      tags:
      - clients
    get:
      parameters:
      - description: Tenant ID
        in: path
        name: tid
        required: true
        type: string
      - description: Client ID
        in: path
        name: cid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/client.ClientResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Error'
      security:
      - AdminAuth: []
      summary: Get client
      tags:
      - clients
    put:
      consumes:
      - application/json
      parameters:
      - description: Tenant ID
        in: path
        name: tid
        required: true
        type: string
      - description: Client ID
        in: path
        name: cid
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/client.UpdateClientRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/client.ClientResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Error'
      security:
      - AdminAuth: []
      summary: Update client
      tags:
      - clients
  /tenants/{tid}/clients/{cid}/api-key:
    post:
      description: Generate a non-expiring API key for a client. The raw key is returned
        once and cannot be retrieved again.
      parameters:
      - description: Tenant ID
        in: path
        name: tid
        required: true
        type: string
      - description: Client ID
        in: path
        name: cid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Error'
      security:
      - AdminAuth: []
      summary: Generate API key
      tags:
      - clients
...
swagger: "2.0"